LQF tag added, modular process assembly prototype implemented (intermediate state)
A fairly significant change in direction for the under-the-hood
implementation but one which starts to allow a very convenient style
for defining and chaining read aware processes.

Of more pressing importance, the LQF flag now works and does what one
would expect in theory. Needs proper verification.

Next all this new stuff needs to be wired up to the CLI