New min_non_edge_reads parameter; slight infrastructure changes to suit
New parameter min_non_edge_reads which extends ADF to check if there are
> min_non_edge_reads, and if not, flag ADF. This also gave me a bit of
momentum to finally start to reconsider FlagResult. "Codes" are now
InfoFlags, which has annoying nomenclature overlap with a Variant Flag
I know. But the point is to standardise all additional info as a
bit set, i.e. a set of flags. ADF now gives you a lot of info in that field
about why a variant is ADF for example - but of course you need to decompose
the information, which is currently tricky and undocumented. Will address soon.

Other minor infrastructure tightening